- (2) changed @PostMapping to @PutMapping; - (3) swapped the logic between model and TO.
